+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+- 2020-03-22
+    - Added new Mile High site: Bi Empire
+    - Fix PornDoe sites
 - 2020-02-09
     - Added new XEmpire site: AllBlackX
 - 2019-12-04